pyproject.toml
[project]
name = "comfyui-subjectbackgroundmotion"
description = "Nodes for subject background motion"
version = "1.0.5"
license = {file = "LICENSE"}
dependencies = ["librosa", "numpy", "pillow>=10.3.0", "scipy", "color-matcher", "matplotlib", "huggingface_hub", "peft"]

[project.urls]
Repository = "https://github.com/Pablerdo/ComfyUI-SubjectBackgroundMotion"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "Pablerdo"
DisplayName = "ComfyUI-SubjectBackgroundMotion"
Icon = ""
